+++ bee-message/CHANGELOG.md
@@ -21,7 +21,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security -->
 
-## 0.1.5 - 2021-05-12
+## 0.1.6 - 2021-12-07
+
+### Changed
+
+ - Update dependencies;
+
+## 0.1.5 - 2021-12-05
 
 ### Removed